refactor(oci): Migrate `init` method to own file
