+++ oci/oci.go
@@ -4,11 +4,6 @@
 // You may not use this file except in compliance with the License.
 package oci
 
-import (
-	"kraftkit.sh/cmdfactory"
-	"kraftkit.sh/packmanager"
-)
-
 var (
 	flagTag           string
 	flagUseMediaTypes bool
@@ -18,35 +13,3 @@ const (
 	DefaultRegistry  = "unikraft.org"
 	DefaultNamespace = "default"
 )
-
-// FIXME(antoineco): avoid init, initialize things where needed
-func init() {
-	// Register a new pkg.Package type
-	_ = packmanager.RegisterPackageManager(
-		OCIFormat,
-		NewOCIManager,
-		WithDefaultAuth(),
-		WithDefaultRegistries(),
-		WithDetectHandler(),
-	)
-
-	// Register additional command-line flags
-	cmdfactory.RegisterFlag(
-		"kraft pkg",
-		cmdfactory.StringVar(
-			&flagTag,
-			"oci-tag",
-			"",
-			"Set the OCI image tag.",
-		),
-	)
-	cmdfactory.RegisterFlag(
-		"kraft pkg",
-		cmdfactory.BoolVar(
-			&flagUseMediaTypes,
-			"oci-use-media-types",
-			false,
-			"Use media types as opposed to well-known paths (experimental).",
-		),
-	)
-}
